Hack:  Trap nil res.Row and raise an error so higher layers can retry.
This happens occasionally but regularly and w/o this hack triggers
a panic when res.Row.Families is dereferenced.  I don't think this
seems normal but haven't been able to identify the root cause.
However, retrying seems to be correct because I'm not seeing
duplicate writes after this error is caught + retried.